Update 3.1
Created a new feature: Preferences

You are now able to select preferences for your profil in order to make these information available to other users on a Lan.

Also there are a lot of bugfixes, code improvements and changes. Most of the are not listed in the projects due to changes, which where required for adding the preference feature.

Closed tickets: Fix #55, Fix #50, Fix #48, Fix #47, Fix #45, Fix #46, Fix #38